refactor MavenDependencyVersionFetcher and add file support
