+++ plugins/core/src/main/kotlin/de/fayard/refreshVersions/core/internal/MavenDependencyVersionsFetcher.kt
@@ -3,46 +3,21 @@ package de.fayard.refreshVersions.core.internal
 import de.fayard.refreshVersions.core.DependencyVersionsFetcher
 import de.fayard.refreshVersions.core.ModuleId
 import de.fayard.refreshVersions.core.Version
-import de.fayard.refreshVersions.core.extensions.okhttp.await
-import okhttp3.OkHttpClient
-import okhttp3.Request
-import retrofit2.HttpException
-import retrofit2.Response
 import java.text.SimpleDateFormat
 import java.util.*
 
-internal class MavenDependencyVersionsFetcher(
-    private val httpClient: OkHttpClient,
+internal abstract class MavenDependencyVersionsFetcher(
     moduleId: ModuleId,
-    repoUrl: String,
-    repoAuthorization: String?
+    repoUrl: String
 ) : DependencyVersionsFetcher(
     moduleId = moduleId,
     repoKey = repoUrl
 ) {
-
-    init {
-        require(repoUrl.endsWith('/'))
-    }
-
-    private val request = Request.Builder().apply {
-        val metadataUrlForArtifact = moduleId.let { (group, name) ->
-            "$repoUrl${group!!.replace('.', '/')}/$name/maven-metadata.xml"
-        }
-        url(OkHttpFileUrlHandler.rewrittenIfFileSchemeUrl(metadataUrlForArtifact))
-        header(
-            name = "Authorization",
-            value = repoAuthorization ?: return@apply
-        )
-    }.build()
+    abstract suspend fun getMetadata(): String
 
     override suspend fun getAvailableVersions(versionFilter: ((Version) -> Boolean)?): SuccessfulResult {
 
-        val xml = httpClient.newCall(request).await().use { response ->
-            if (response.isSuccessful) {
-                response.use { it.body!!.string() }
-            } else throw HttpException(Response.error<Any?>(response.code, response.body!!))
-        }
+        val xml = getMetadata()
 
         val allVersions = parseVersionsFromMavenMetaData(xml)
         return SuccessfulResult(
