+++ plugins/core/src/main/kotlin/de/fayard/refreshVersions/core/internal/MavenDependencyVersionsFetcherFILE.kt
@@ -1,0 +1,22 @@
+package de.fayard.refreshVersions.core.internal
+
+import de.fayard.refreshVersions.core.ModuleId
+import java.io.File
+
+internal class MavenDependencyVersionsFetcherFILE(
+    moduleId: ModuleId,
+    repoUrl: String
+) : MavenDependencyVersionsFetcher(
+    moduleId = moduleId,
+    repoUrl = repoUrl
+) {
+    private val baseFolder = File(repoUrl)
+
+    init {
+//        require(repoUrl.endsWith('/'))
+    }
+
+    override suspend fun getMetadata(): String {
+        return baseFolder.resolve("${moduleId.group!!.replace('.', '/')}/${moduleId.name}/maven-metadata.xml").readText()
+    }
+}
